requirements.txt
Django>=3.0,<4.0
djangorestframework==3.13.1
psycopg2-binary>=2.8
markdown==3.3.6
django-filter==21.1
# Dev
pytest==7.0.0
pytest-django==4.5.2
pytest-cov==3.0.0
# Swagger
pyyaml==6.0
django-rest-swagger==2.2.0
# Auth token
djangorestframework-simplejwt==5.1.0


